Fold Higgs streaming vocoder back into stages.py
Drop the separate streaming_vocoder module and keep the vocoder code
co-located with the rest of the Higgs TTS pipeline factories. The
original _vocode closure is preserved as the non-streaming path; the
new HiggsVocoderScheduler wraps it for the streaming case.

Also: drop the unused outbox ctor parameter on HiggsTTSModelRunner,
the hasattr() guard on a freshly-constructed runner, the always-True
"stream" metadata key, the write-only _HiggsStreamState.audio_chunks
field, and the redundant .detach().cpu() calls on tensors that are
already detached and on host.
